requirements.txt
discord.py==2.3.2
gallery-dl==1.26.6
yt_dlp==2025.2.19